+++ modulo3/astro-match/src/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom/cjs/react-router-dom.min';
 import App from './App';
 
 ReactDOM.render(
-	<React.StrictMode>
+	<BrowserRouter>
 		<App />
-	</React.StrictMode>,
+	</BrowserRouter>,
 	document.getElementById('root')
 );
